tickets/src/routes/new.ts
import { requireAuth, validateRequest } from "@v0x-shared/common";

import express, { Request, Response } from "express";
import { body } from "express-validator";
import { TicketCreatedPublisher } from "../events/publishers/ticket-created-publisher";
import { Ticket } from "../models/Ticket";
import { natsWrapper } from "../nats-wrapper";

const router = express.Router();

router.post(
  "/api/tickets",
  requireAuth,
  [
    body("title")
      .not()
      .isEmpty()
      .withMessage("Ticket name is required.")
      .isLength({ min: 3, max: 20 })
      .withMessage("Your ticket name need to be between 3 and 20 characters."),
    body("price")
      .trim()
      .isFloat({ gt: 0.5, lt: 10000 })
      .withMessage("Only numbers can be provided for price"),
  ],
  validateRequest,
  async (req: Request, res: Response) => {
    const { title, price } = req.body;

    const ticket = Ticket.build({
      title,
      price,
      userId: req.currentUser!.id,
    });

    await ticket.save();
    await new TicketCreatedPublisher(natsWrapper.client).publish({
      id: ticket.id,
      title: ticket.title,
      price: ticket.price,
      userId: ticket.userId,
    });

    res.status(201).send(ticket);
  }
);

export { router as createTicketRouter };
